refactor(mvc): cleanse package structures
