+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,8 +4,6 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.tobe.HandlerAdapterRegistry;
-import nextstep.mvc.controller.tobe.HandlerMappingRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
